+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Moved to Rust 2018.
+- `xmlparser` updated.
 
 ## [0.8.0] - 2019-12-21
 ### Added