+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.9.1] - 2020-02-09
 ### Changed
 - A better entity loop detection. A document can have an unlimited
   number of references at zero depth now.
@@ -113,7 +115,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `ExpandedName::has_namespace`. `ExpandedName::namespace` will return `Option` now.
 - `Error::NestedEntityReference`.
 
-[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.9.1...HEAD
+[0.9.1]: https://github.com/RazrFalcon/roxmltree/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.3...v0.8.0
 [0.7.3]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.2...v0.7.3